Split core logging logic into a dedicated class
The logic in AbstractLogService is now generalized in a DefaultLogger
implementation which implements Logger. The AbstractLogService now
simply keeps an instance of DefaultLogger to which it delegates
everything, apart from calls to alwaysLog, which is marks abstract.

To better understand the motivation for this change, see next commits.
